+++ src/inu/const.py
@@ -1,4 +1,4 @@
-INU_BUILD = 32
+INU_BUILD = 33
 
 
 class LogLevel: